+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- _SKU Selector_ item with false value as class name.
+
 ## [1.5.0] - 2018-6-11
 ### Added
 - Max height of the logo image
@@ -19,7 +22,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - _SKU Selector_ item's discount badge position
 - Put the _ProductPrice_ schema inside it's Component.
 - README typos.
-- _SKU Selector_ item with false value as class name.
 
 ## [1.4.0] - 2018-6-6
 ### Added